Fixed [Devil May Cry 5 Controller Overlay] Theme
Was a byproduct of the "Redesigned New DualShock 4 Theme", but now fixed. (At last, the storm is no longer approaching!)

This is due to subimages files connected to the Default Theme, specifically: the Base, Triggers and Analog Sticks. Because of that, the entire theme is misaligned and base active images are stretched.

To fix this; I basically redid the entire theme (and updated my old workflow file) using updated Default .json as a basis. Now the DMC5 Theme is a complete parity with the Default Theme.