New Steam Controller Theme
This theme is based on the official lineart released by Valve, while I change the .json file to be closer to PlayStation Themes. This should make things easier to transfer between Current and Future Themes.

Alongside it, the old Theme will be kept around for Backwards Compatibility reasons. As of right now, themes that uses the Shared Assets features may not work. I will be fixing this later down in the line