Added New Steam Controller's Compact theme, Changed Winform.Interpolation for both DS4 and SC Themes.
Added the Compact Theme to the New Steam Controller Theme. This should be closer to the Classic SC Theme. This also includes "All" .json that includes Motion Sensors (Gyro and Accel)

I also adjusted the Winform.Interpolation for both DS4 and Steam Controller themes as the Base Controller Images are already high quality as it is.